Extract one-shot initiator helpers from peer_link_client.py
Convert the 1881-line peer_link_client.py monolith into a
peer_link_client/ package. PR 1 of the split: move the top-half
one-shot WS round-trip helpers (drive_initiator_round_trip,
preview_pair, request_pair, await_pair_status) to a new
one_shot.py submodule; PeerLinkClient stays in client.py.

__init__.py re-exports the public surface (PeerLinkClient + the
one-shot funcs + the *Error / *Result types from _client_models)
plus the underscore-prefixed names tests depend on. Existing
'from .peer_link_client import X' callers keep working through
the package re-export.

Test patches that monkeypatched module attributes are redirected
to the submodule the symbol now lives in (e.g.
peer_link_client.client.run_peer_link_heartbeat) so the
production-side reads still see the override. No behaviour
change to production code paths.
